Fix minor mistakes and cleanup in tests
Tests are now sorted according to their real counterparts.
Log doesn't get spammed by HTML during tests anymore.
SeasonServiceTest uses dummies now instead of replicating the functions
in SeasonService.
Remove duplicate TrackedStatisticsDummy.
