+++ src/test/java/de/adesso/kicker/events/MatchVerifiedEventDummy.java
@@ -1,6 +1,6 @@
 package de.adesso.kicker.events;
 
-import de.adesso.kicker.match.MatchDummy;
+import de.adesso.kicker.match.persistence.MatchDummy;
 import de.adesso.kicker.match.service.events.MatchVerifiedEvent;
 
 public class MatchVerifiedEventDummy {
