+++ src/test/resources/application-test.properties
@@ -3,3 +3,5 @@ keycloak.realm=test
 keycloak.resource=kicker
 keycloak.public-client=true
 keycloak.principal-attribute=preferred_username
+logging.level.root=OFF
+spring.main.banner-mode=off
